requirements.txt
PyPDF2==1.26.0
PyYAML==3.11
